+++ src/index_buffer.hpp
@@ -9,11 +9,11 @@ class IndexBufferHandle {
 public:
     IndexBufferHandle();
 
-    IndexBufferHandle(uint32_t value);
+    IndexBufferHandle(uint64_t value);
 
-    void set(uint32_t value);
+    void set(uint64_t value);
 
-    [[nodiscard]] uint32_t value() const;
+    [[nodiscard]] uint64_t value() const;
 
     [[nodiscard]] bool operator==(const IndexBufferHandle& other) const;
 
@@ -21,7 +21,7 @@ public:
 
 private:
     bool m_initialized = false;
-    uint32_t m_value;
+    uint64_t m_value;
 };
 
 class Renderer;
@@ -62,7 +62,7 @@ template <>
 struct hash<mve::IndexBufferHandle> {
     std::size_t operator()(const mve::IndexBufferHandle& handle) const
     {
-        return hash<uint32_t>()(handle.value());
+        return hash<uint64_t>()(handle.value());
     }
 };
