+++ src/vertex_buffer.hpp
@@ -9,11 +9,11 @@ class VertexBufferHandle {
 public:
     VertexBufferHandle();
 
-    VertexBufferHandle(uint32_t value);
+    VertexBufferHandle(uint64_t value);
 
-    void set(uint32_t value);
+    void set(uint64_t value);
 
-    [[nodiscard]] uint32_t value() const;
+    [[nodiscard]] uint64_t value() const;
 
     [[nodiscard]] bool operator==(const VertexBufferHandle& other) const;
 
@@ -21,7 +21,7 @@ public:
 
 private:
     bool m_initialized = false;
-    uint32_t m_value;
+    uint64_t m_value;
 };
 
 class Renderer;
@@ -64,7 +64,7 @@ template <>
 struct hash<mve::VertexBufferHandle> {
     std::size_t operator()(const mve::VertexBufferHandle& handle) const
     {
-        return hash<uint32_t>()(handle.value());
+        return hash<uint64_t>()(handle.value());
     }
 };
